+++ packages/components/src/Button/IconButton.tsx
@@ -128,6 +128,7 @@ export const IconButtonStyle = styled.button.withConfig({
 const IconButtonComponent = forwardRef(
   (props: IconButtonProps, forwardRef: Ref<HTMLButtonElement>) => {
     const {
+      'aria-expanded': ariaExpanded,
       className,
       icon,
       id,
@@ -164,7 +165,8 @@ const IconButtonComponent = forwardRef(
       tooltip,
     } = useTooltip({
       content: label,
-      disabled: tooltipDisabled || hasOuterTooltip,
+      // ariaExpanded = true indicates an open Popover or Menu – don't show the tooltip
+      disabled: tooltipDisabled || hasOuterTooltip || ariaExpanded === true,
       id: id ? `${id}-tooltip` : undefined,
       placement: tooltipPlacement,
       textAlign: tooltipTextAlign,
@@ -183,6 +185,7 @@ const IconButtonComponent = forwardRef(
     return (
       <ButtonBase
         aria-describedby={ariaDescribedBy}
+        aria-expanded={ariaExpanded}
         aria-pressed={toggle ? true : undefined}
         ref={actualRef}
         p="none"