+++ packages/components/src/DataTable/Item/DataTableItem.tsx
@@ -28,7 +28,7 @@ import { CompatibleHTMLProps } from '@looker/design-tokens'
 import styled from 'styled-components'
 import React, { FC, ReactNode, useContext, useRef } from 'react'
 import { IconButton } from '../../Button'
-import { Menu, MenuDisclosure, MenuList } from '../../Menu'
+import { Menu } from '../../Menu'
 import { DataTableContext } from '../DataTableContext'
 import { DataTableRow } from './DataTableRow'
 import { ItemTarget, ItemTargetGroup } from './ItemTarget'
@@ -94,11 +94,8 @@ const DataTableItemLayout: FC<DataTableItemProps> = ({
       {actionPrimary && <ItemTarget>{actionPrimary}</ItemTarget>}
       {actions && (
         <ItemTarget>
-          <Menu>
-            <MenuDisclosure tooltip={actionsTooltip}>
-              <IconButton icon="DotsVert" size="small" label={actionsTooltip} />
-            </MenuDisclosure>
-            <MenuList>{actions}</MenuList>
+          <Menu content={actions}>
+            <IconButton icon="DotsVert" size="small" label={actionsTooltip} />
           </Menu>
         </ItemTarget>
       )}